Move tribute webhook route to service
